requirements.txt
asn1crypto==1.2.0
atomicwrites==1.4.0
attrs==19.3.0
aws-sam-translator==1.22.0
aws-xray-sdk==2.5.0
bcrypt==3.1.7
boto==2.49.0
boto3==1.9.250
botocore==1.12.250
certifi==2019.9.11
cffi==1.13.0
cfn-lint==0.30.1
chardet==3.0.4
cryptography==2.7
decorator==4.4.2
docker==4.2.0
docutils==0.15.2
ecdsa==0.15
future==0.18.2
idna==2.8
importlib-metadata==1.6.0
importlib-resources==1.5.0
Jinja2==2.11.2
jmespath==0.9.4
jsondiff==1.1.2
jsonpatch==1.25
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
lxml==4.4.1
MarkupSafe==1.1.1
mock==4.0.2
more-itertools==8.2.0
moto==1.3.14
networkx==2.4
numpy==1.17.2
oauthlib==3.1.0
packaging==20.3
pandas==0.25.1
paramiko==2.6.0
pathlib==1.0.1
pluggy==0.13.1
py==1.8.1
pyasn1==0.4.8
pycparser==2.19
pymarc==3.1.13
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.2.4
pytest-mock==3.1.0
python-dateutil==2.8.0
python-jose==3.1.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-mock==1.8.0
requests-oauthlib==1.2.0
responses==0.10.14
rsa==4.0
s3transfer==0.2.1
six==1.12.0
sshpubkeys==3.1.0
urllib3==1.25.6
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
zope.interface==4.6.0
